docs: detail virtual pin configuration
